+++ lib/xdot/xdot.h
@@ -60,7 +60,7 @@ typedef struct {
 typedef struct {
   double x0, y0, r0;
   double x1, y1, r1;
-  int n_stops;
+  unsigned n_stops;
   xdot_color_stop *stops;
 } xdot_radial_grad;